+++ apps/src/core/bitmaps/CellBitmap.cpp
@@ -2,22 +2,22 @@
 
 namespace DirtSim {
 
-CellBitmap::CellBitmap(uint32_t width, uint32_t height) : grid_width_(width), grid_height_(height)
+CellBitmap::CellBitmap(int width, int height)
+    : grid_width_(static_cast<int16_t>(width)), grid_height_(static_cast<int16_t>(height))
 {
     // Calculate number of 8×8 blocks needed (round up for partial blocks).
-    blocks_x_ = (width + BLOCK_SIZE - 1) / BLOCK_SIZE;
-    blocks_y_ = (height + BLOCK_SIZE - 1) / BLOCK_SIZE;
+    blocks_x_ = static_cast<int16_t>((width + BLOCK_SIZE - 1) / BLOCK_SIZE);
+    blocks_y_ = static_cast<int16_t>((height + BLOCK_SIZE - 1) / BLOCK_SIZE);
 
     // Allocate one uint64_t per 8×8 block, initialized to zero.
-    blocks_.resize(blocks_x_ * blocks_y_, 0);
+    blocks_.resize(static_cast<size_t>(blocks_x_ * blocks_y_), 0);
 }
 
-inline void CellBitmap::cellToBlockAndBit(
-    uint32_t x, uint32_t y, uint32_t& block_idx, int& bit_idx) const
+inline void CellBitmap::cellToBlockAndBit(int x, int y, int& block_idx, int& bit_idx) const
 {
     // Block coordinates using bit shifts instead of division.
-    uint32_t block_x = x >> 3; // x / 8
-    uint32_t block_y = y >> 3; // y / 8
+    int block_x = x >> 3; // x / 8
+    int block_y = y >> 3; // y / 8
     block_idx = block_y * blocks_x_ + block_x;
 
     // Local coordinates within the 8×8 block using bit masks instead of modulo.
@@ -28,50 +28,50 @@ inline void CellBitmap::cellToBlockAndBit(
     bit_idx = (local_y << 3) | local_x; // local_y * 8 + local_x
 }
 
-void CellBitmap::set(uint32_t x, uint32_t y)
+void CellBitmap::set(int x, int y)
 {
-    uint32_t block_idx;
+    int block_idx;
     int bit_idx;
     cellToBlockAndBit(x, y, block_idx, bit_idx);
     blocks_[block_idx] |= (1ULL << bit_idx);
 }
 
-void CellBitmap::clear(uint32_t x, uint32_t y)
+void CellBitmap::clear(int x, int y)
 {
-    uint32_t block_idx;
+    int block_idx;
     int bit_idx;
     cellToBlockAndBit(x, y, block_idx, bit_idx);
     blocks_[block_idx] &= ~(1ULL << bit_idx);
 }
 
-bool CellBitmap::isSet(uint32_t x, uint32_t y) const
+bool CellBitmap::isSet(int x, int y) const
 {
-    uint32_t block_idx;
+    int block_idx;
     int bit_idx;
     cellToBlockAndBit(x, y, block_idx, bit_idx);
     return (blocks_[block_idx] >> bit_idx) & 1;
 }
 
-uint64_t CellBitmap::getBlock(uint32_t block_x, uint32_t block_y) const
+uint64_t CellBitmap::getBlock(int block_x, int block_y) const
 {
-    uint32_t block_idx = block_y * blocks_x_ + block_x;
+    int block_idx = block_y * blocks_x_ + block_x;
     return blocks_[block_idx];
 }
 
-bool CellBitmap::isBlockAllSet(uint32_t block_x, uint32_t block_y) const
+bool CellBitmap::isBlockAllSet(int block_x, int block_y) const
 {
     return getBlock(block_x, block_y) == 0xFFFFFFFFFFFFFFFFULL;
 }
 
-bool CellBitmap::isBlockAllClear(uint32_t block_x, uint32_t block_y) const
+bool CellBitmap::isBlockAllClear(int block_x, int block_y) const
 {
     return getBlock(block_x, block_y) == 0;
 }
 
-Neighborhood3x3 CellBitmap::getNeighborhood3x3(uint32_t x, uint32_t y) const
+Neighborhood3x3 CellBitmap::getNeighborhood3x3(int x, int y) const
 {
-    uint32_t block_x = x >> 3;
-    uint32_t block_y = y >> 3;
+    int block_x = x >> 3;
+    int block_y = y >> 3;
     int local_x = x & 7;
     int local_y = y & 7;
 
@@ -104,12 +104,11 @@ Neighborhood3x3 CellBitmap::getNeighborhood3x3(uint32_t x, uint32_t y) const
     for (int dy = -1; dy <= 1; ++dy) {
         for (int dx = -1; dx <= 1; ++dx) {
             int bit_pos = (dy + 1) * 3 + (dx + 1);
-            int nx = static_cast<int>(x) + dx;
-            int ny = static_cast<int>(y) + dy;
+            int nx = x + dx;
+            int ny = y + dy;
 
             // Check grid boundaries.
-            if (nx >= 0 && nx < static_cast<int>(grid_width_) && ny >= 0
-                && ny < static_cast<int>(grid_height_)) {
+            if (nx >= 0 && nx < grid_width_ && ny >= 0 && ny < grid_height_) {
                 // Set validity bit.
                 result |= (1ULL << (9 + bit_pos));
